+++ src/view/startpage/PopupComponent.tsx
@@ -1,8 +1,14 @@
 interface PopupComponentProps {
   // Props
-  // TODO
+  message: String;
+  messageType: string;
 }
 
-export default function PopupComponent({}: PopupComponentProps) {
-  return <div>Popup Component</div>;
+export default function PopupComponent({message, messageType}: PopupComponentProps) {
+  return (
+  <div className="alert alert-primary" role="alert">
+    <b>{messageType}</b><br></br>
+    {message}
+  </div>
+  );
 }